backend/requirements.txt
django>=4.2.20
djangorestframework>=3.15.2
pinecone>=6.0.2
langchain>=0.3.20
openai>=1.66.3
python-dotenv>=1.0.1 